Survey/requirements.txt

click==8.1.3
elastic-transport==8.1.2
elasticsearch==8.2.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
